Compiled third_pixels_post.c in LST
git-svn-id: https://edclxs131.cr.usgs.gov:18080/svn/lsrd/lst/trunk@916 48de96a7-6a0d-487c-b84a-40161698abc1

+++ not-validated-prototype-l5-7_lst/src/second_narr.c
@@ -494,7 +494,7 @@ int second_narr
     int num_points,             /*I: number of narr points */
     float alb,                  /*I: albedo */ 
     char **case_list,           /*I: modtran run list */
-    double **results,           /*O: atmospheric parameter for modtarn run */
+    float **results,            /*O: atmospheric parameter for modtarn run */
     bool verbose                /*I: value to indicate if intermediate messages 
                                      be printed */
 )
